refactor(number generator): number_generator 패키지로 이동
