+++ src/main/java/lotto/number_generator/TestNumberGenerator.java
@@ -1,10 +1,10 @@
-package lotto;
+package lotto.number_generator;
 
 import java.util.List;
 
 public class TestNumberGenerator implements NumberGenerator{
     @Override
     public List<Integer> generate() {
-        return List.of(1, 2, 3, 4, 5, 6);
+        return List.of(6, 5, 4, 3, 2, 1);
     }
 }
